+++ newsmth/more/SMConfig.m
@@ -69,7 +69,7 @@
         boards = [[NSMutableArray alloc] init];
     }
     [boards insertObject:dict atIndex:0];
-    for (int i = boards.count - 1; i > 0; --i) {
+    for (int i = (int)boards.count - 1; i > 0; --i) {
         NSDictionary *b = boards[i];
         if (i > 10 || [b[@"name"] isEqualToString:board.name]) {    // 最多保存10个最近浏览版面
             [boards removeObjectAtIndex:i];
